pyproject.toml
[project]
name = "bcollar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.3",
]
